+++ lib/src/template.dart
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// ignore: deprecated_member_use
 import 'dart:cli';
 import 'dart:io';
 
@@ -28,6 +29,7 @@ final _cache = p.PathMap<String>();
 /// Note: This function only supports simple variable replacement. It does not
 /// support any additional Mustache features.
 String renderTemplate(String path, Map<String, String> variables) {
+  // ignore: deprecated_member_use
   path = p.join(waitFor(cliPkgSrc), 'templates', path);
   var text =
       _cache.putIfAbsent(path, () => File("$path.mustache").readAsStringSync());